const armageddon

+++ include/mac/GSymLookup.h
@@ -120,14 +120,14 @@ private:
 		if (info->dli_sname) {
 			symbol = info->dli_sname;
 		} else {
-			snprintf(symbuf, sizeof(symbuf), "0x%x", info->dli_saddr);
+			snprintf(symbuf, sizeof(symbuf), "0x%x", (int)info->dli_saddr);
 		}
 		
 		return snprintf(buf, size,
 				_BACKTRACE_FORMAT,
 				frame,
 				image,
-				addr,
+				(unsigned) addr,
 				symbol,
 				(char*)addr - (char*)info->dli_saddr) + 1;
 	}
